# shared-schema/queueserver_service.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Bluesky HTTP Server",
    "description": "REST + WebSocket surface of the Bluesky Queue Server. Drives an experiment from\nthree angles: **queue** (plans and instructions waiting to run), **Run Engine**\n(control over the currently-running plan), and **worker environment** (startup\nscripts, plans, devices, and tasks inside the worker process).\n\nAll endpoints live under `/api`. Clients authenticate with a bearer token\n(access or API key); anonymous access, when enabled, is scoped to `read:status`\nonly. Most endpoints return `{\"success\": bool, \"msg\": str, ...}` envelopes \u2014\ninspect `success` on each response.\n\n### WebSocket endpoints (not listed below)\n\nThree push streams are available; they do not render in Swagger UI because\nOpenAPI has no WebSocket concept:\n\n- `GET /api/status/ws` \u2014 pushes the `/api/status` payload as it changes\n  (scope `read:monitor`).\n- `GET /api/info/ws` \u2014 pushes system-info snapshots (scope `read:monitor`).\n- `GET /api/console_output/ws` \u2014 pushes captured worker stdout/stderr\n  (scope `read:console`).\n\nClients authenticate WebSockets by passing the bearer token via a cookie or\nthe `Authorization` header on the upgrade request.\n",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/manager/stop": {
      "post": {
        "tags": [
          "Manager"
        ],
        "summary": "Stop the RE Manager",
        "description": "Stop RE Manager. Unlike crash-and-restart behaviour, the manager will NOT be auto-restarted by the watchdog after a stop issued via this endpoint. Required scope: `write:manager:stop`.",
        "operationId": "manager_stop_handler_api_manager_stop_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "type": "object",
                "title": "Payload",
                "default": {}
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessMsgResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": [
              "write:manager:stop"
            ]
          },
          {
            "APIKeyQuery": [
              "write:manager:stop"
            ]
          },
          {
            "APIKeyAuthorizationHeader": [
              "write:manager:stop"
            ]
          },
          {
            "APIKeyCookie": [
              "write:manager:stop"
            ]
          }
        ]
      }
    },
    "/api/test/manager/kill": {
      "post": {
        "tags": [
          "Testing"
        ],
        "summary": "Kill the manager event loop (testing only)",
        "description": "Halt the manager event loop to test client-side timeout handling and watchdog restart behaviour. Not for production use. Required scope: `write:testing`.",
        "operationId": "test_manager_kill_handler_api_test_manager_kill_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessMsgResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": [
              "write:testing"
            ]
          },
          {
            "APIKeyQuery": [
              "write:testing"
            ]
          },
          {
            "APIKeyAuthorizationHeader": [
              "write:testing"
            ]
          },
          {
            "APIKeyCookie": [
              "write:testing"
            ]
          }
        ]
      }
    },
    "/api/test/server/sleep": {
      "get": {
        "tags": [
          "Testing"
        ],
        "summary": "Sleep on the server (testing only)",
        "description": "Sleep for `time` seconds then return success. Does not block the event loop or manager calls. Used to exercise client timeout handling. Required scope: `read:testing`.",
        "operationId": "test_server_sleep_handler_api_test_server_sleep_get",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "type": "object",
                "title": "Payload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessMsgResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": [
              "read:testing"
            ]
          },
          {
            "APIKeyQuery": [
              "read:testing"
            ]
          },
          {
            "APIKeyAuthorizationHeader": [
              "read:testing"
            ]
          },
          {
            "APIKeyCookie": [
              "read:testing"
            ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SuccessMsgResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "msg": {
            "type": "string",
            "title": "Msg"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "success",
          "msg"
        ],
        "title": "SuccessMsgResponse",
        "description": "Bare ``{success, msg}`` envelope shared by control routes that return no payload."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "PLACEHOLDER",
            "refreshUrl": "token/refresh"
          }
        }
      },
      "APIKeyQuery": {
        "type": "apiKey",
        "in": "query",
        "name": "api_key"
      },
      "APIKeyAuthorizationHeader": {
        "type": "apiKey",
        "description": "Prefix value with 'Apikey ' as in, 'Apikey SECRET'",
        "in": "header",
        "name": "Authorization"
      },
      "APIKeyCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "bluesky_httpserver_api_key"
      }
    }
  },
...
}